pyproject.toml
[project]
name = "productivity-agent"
version = "0.1.0"
description = "Personal Productivity Agent - MEST Week 2"
requires-python = ">=3.13"
dependencies = [
    "pydantic-ai[gemini]>=0.0.14",
    "streamlit>=1.40.0",
    "python-dotenv>=1.0.0",
    "tavily-python>=0.3.0",
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "aiosqlite>=0.20.0",
    "google-auth-oauthlib>=1.3.0",
    "google-api-python-client>=2.192.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent", "mcp_server"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]
